feat: refactor library structure
