+++ src/main/java/com/github/playerforcehd/gcaptchavalidator/ValidationResponse.java
@@ -1,15 +1,42 @@
-package com.github.playerforcehd.gcaptchavalidator.model.response;
+/*
+ * The MIT License
+ *
+ * Copyright (c) 2020 Pascal Zarrad
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+
+package com.github.playerforcehd.gcaptchavalidator;
+
+import com.github.playerforcehd.gcaptchavalidator.param.ClientType;
+import com.github.playerforcehd.gcaptchavalidator.param.ValidationError;
 
 import java.util.Date;
 import java.util.Map;
 
 /**
- * An immutable implementation of the {@link ValidationResponse}.
+ * An immutable implementation of the {@link CaptchaValidationResponse}.
  *
  * @author Pascal Zarrad
  * @since 3.0.0
  */
-public final class ImmutableResponse implements ValidationResponse {
+public final class ValidationResponse implements CaptchaValidationResponse {
     /**
      * The state if the response was successful
      */
@@ -50,7 +77,7 @@ public final class ImmutableResponse implements ValidationResponse {
      * @param errors The errors returned in the response
      * @param additionalData Additional data that is not defined in the standard response from Google
      */
-    public ImmutableResponse(
+    ValidationResponse(
         boolean succeeded,
         Date challengeTimestamp,
         ClientType clientType,
