+++ src/main/java/com/github/playerforcehd/gcaptchavalidator/ValidatorConfiguration.java
@@ -1,6 +1,30 @@
-package com.github.playerforcehd.gcaptchavalidator.model.validator;
+/*
+ * The MIT License
+ *
+ * Copyright (c) 2020 Pascal Zarrad
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
 
-import com.github.playerforcehd.gcaptchavalidator.model.ReCaptchaVersion;
+package com.github.playerforcehd.gcaptchavalidator;
+
+import com.github.playerforcehd.gcaptchavalidator.param.ReCaptchaVersion;
 
 import java.util.Map;
 
@@ -13,9 +37,9 @@ import java.util.Map;
  *
  * @author Pascal Zarrad
  * @since 3.0.0
- * @see com.github.playerforcehd.gcaptchavalidator.model.validator.ValidatorConfiguration
+ * @see CaptchaValidatorConfiguration
  */
-public class MutableValidatorConfiguration implements ValidatorConfiguration {
+public class ValidatorConfiguration implements CaptchaValidatorConfiguration {
     /**
      * The ReCaptcha server side secret required for validation.
      */
@@ -46,7 +70,7 @@ public class MutableValidatorConfiguration implements ValidatorConfiguration {
      * @param verifierUrl The URL where the requests will be send to
      * @param httpHeaders The HTTP headers to send with requests using this configuration
      */
-    public MutableValidatorConfiguration(
+    public ValidatorConfiguration(
         String secretToken,
         ReCaptchaVersion reCaptchaVersion,
         String verifierUrl,
@@ -64,7 +88,7 @@ public class MutableValidatorConfiguration implements ValidatorConfiguration {
     }
 
     @Override
-    public MutableValidatorConfiguration setSecretToken(String secretToken) {
+    public ValidatorConfiguration setSecretToken(String secretToken) {
         this.secretToken = secretToken;
 
         return  this;
@@ -76,7 +100,7 @@ public class MutableValidatorConfiguration implements ValidatorConfiguration {
     }
 
     @Override
-    public MutableValidatorConfiguration setReCaptchaVersion(ReCaptchaVersion reCaptchaVersion) {
+    public ValidatorConfiguration setReCaptchaVersion(ReCaptchaVersion reCaptchaVersion) {
         this.reCaptchaVersion = reCaptchaVersion;
 
         return this;
@@ -88,7 +112,7 @@ public class MutableValidatorConfiguration implements ValidatorConfiguration {
     }
 
     @Override
-    public MutableValidatorConfiguration setVerifierUrl(String verifierUrl) {
+    public ValidatorConfiguration setVerifierUrl(String verifierUrl) {
         this.verifierUrl = verifierUrl;
 
         return this;
@@ -100,7 +124,7 @@ public class MutableValidatorConfiguration implements ValidatorConfiguration {
     }
 
     @Override
-    public ValidatorConfiguration setHttpHeaders(Map<String, String> httpHeaders) {
+    public CaptchaValidatorConfiguration setHttpHeaders(Map<String, String> httpHeaders) {
         this.httpHeaders = httpHeaders;
 
         return this;
