--- src/main/java/com/github/playerforcehd/gcaptchavalidator/model/ReCaptchaVersion.java
@@ -1,15 +1,0 @@
-package com.github.playerforcehd.gcaptchavalidator.model;
-
-/**
- * List of all ReCaptcha versions which are supported by the library.
- *
- * This allows exact specification of the targeted version to support
- * features that are not available in older/newer versions.
- *
- * @author Pascal Zarrad
- * @since 3.0.0
- */
-public enum ReCaptchaVersion {
-    VERSION_2,
-    VERSION_3;
-}
